Switch over from requirements.txt to Pipfile
This is the new recommended way to install Python packages:

https://www.ostechnix.com/pipenv-officially-recommended-python-packaging-tool/

PIP_USER (aka --user) has been set so the packages will continue to be installed
in /vendor. PIPENV_SYSTEM (aka --system) to use the system install Python, no
need to create a virtual env since this is in a container.

This avoids having to use 'pipenv shell' or 'pipenv run' and
everything runs as usual.